chore: rename jsonFormating ts file
